+++ include/mupdf/pdf/document.h
@@ -671,8 +671,8 @@ void pdf_delete_page_range(fz_context *ctx, pdf_document *doc, int start, int en
 /*
 	Get page label (string) from a page number (index).
 */
-void pdf_page_label(fz_context *ctx, pdf_document *doc, int page, char *buf, int size);
-void pdf_page_label_imp(fz_context *ctx, fz_document *doc, int chapter, int page, char *buf, int size);
+void pdf_page_label(fz_context *ctx, pdf_document *doc, int page, char *buf, size_t size);
+void pdf_page_label_imp(fz_context *ctx, fz_document *doc, int chapter, int page, char *buf, size_t size);
 
 typedef enum {
 	PDF_PAGE_LABEL_NONE = 0,
